backend/requirements.txt

fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.10.3
pydantic-settings==2.6.1

sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
bcrypt==4.2.1
PyJWT==2.8.0

openai==1.55.3


nltk==3.9.1
langdetect==1.0.9
chardet==5.2.0

PyPDF2==3.0.1
python-docx==1.1.2

python-dotenv==1.0.1

pydantic-settings==2.6.1
numpy>=1.24.3

dnspython==2.7.0
email-validator==2.2.0